Fix compile problems on Raspberry Pi (Issue #9)

+++ cups/tls-gnutls.c
@@ -702,7 +702,7 @@ httpLoadCredentials(
         alloc_data += 1024;
       }
 
-      decoded = alloc_data - num_data;
+      decoded = (int)(alloc_data - num_data);
       httpDecode64_2((char *)data + num_data, &decoded, line);
       num_data += (size_t)decoded;
     }
@@ -945,7 +945,7 @@ http_gnutls_load_crl(void)
 	    alloc_data += 1024;
 	  }
 
-	  decoded = alloc_data - num_data;
+	  decoded = (int)(alloc_data - num_data);
 	  httpDecode64_2((char *)data + num_data, &decoded, line);
 	  num_data += (size_t)decoded;
 	}
